==> 01.ipynb <==
# Анализ факторов экологии на результаты ЕГЭ

## ЕГЭ

https://ivo.garant.ru/#/document/406882488

### Досрочный период

22 марта (пятница) — география, литература;
26 марта (вторник) — русский язык;
29 марта (пятница) — ЕГЭ по математике базового уровня, ЕГЭ по математике профильного уровня;
2 апреля (вторник) — биология, иностранные языки (английский, испанский, китайский, немецкий, французский) (письменная часть), физика;
5 апреля (пятница) — иностранные языки (английский, испанский, китайский, немецкий, французский) (устная часть);
9 апреля (вторник) — информатика, обществознание;
12 апреля (пятница) — история, химия.

### Резервные дни

15 апреля (понедельник) — русский язык;
18 апреля (четверг) — ЕГЭ по математике базового уровня, ЕГЭ по математике профильного уровня;
19 апреля (пятница) — биология, иностранные языки (английский, испанский, китайский, немецкий, французский) (письменная часть), литература, обществознание, физика;
22 апреля (понедельник) — география, иностранные языки (английский, испанский, китайский, немецкий, французский) (устная часть), информатика, история, химия;

Основной период

23 мая (четверг) — география, литература, химия;
28 мая (вторник) — русский язык;
29 мая (среда) — русский язык;
31 мая (пятница) — ЕГЭ по математике базового уровня, ЕГЭ по математике профильного уровня;
1 июня (суббота) — ЕГЭ по математике базового уровня, ЕГЭ по математике профильного уровня;
4 июня (вторник) — обществознание, физика;
7 июня (пятница) — иностранные языки (английский, испанский, китайский, немецкий, французский) (устная часть), информатика;
8 июня (суббота) — иностранные языки (английский, испанский, китайский, немецкий, французский) (устная часть), информатика;
11 июня (вторник) — биология, иностранные языки (английский, испанский, китайский, немецкий, французский) (письменная часть), история.


### Дополнительный период

4 сентября (среда) — русский язык;
9 сентября (понедельник) — ЕГЭ по математике базового уровня.
23 сентября (понедельник) — ЕГЭ по математике базового уровня, русский язык.

ЕГЭ начинают сдавать с марта и заканчивают в сентябре

import os
import re
import json
import gzip
import time
import zipfile
import requests
import numpy as np
import pandas as pd
from datetime import date
from decimal import Decimal
#from tokens import YANDEX_ORG
#from tokens import YANDEX_GEO # Раскоменить для работы с API Яндекс.Организации и получения координат
#from tokens import DATA_MOS_RU # Раскоменить для работы с API data.mos.ru
#from tokens import MAPS_CO
MAPS_CO = 0 # # Закоментить строку если есть ключ
YANDEX_GEO = 0 # Закоментить строку если есть ключ
COUNTER = 0 # глобальный счетчик, можно использовать для подсчета кол-ва запросов по АПИ
LIMIT = 1000 # лимит запросов по АПИ и кол-во записей в одно запросе
CACHE = {} # глобальный кэш, в процессе скрейпинга не обновлять!
SETS = './sets/' # папка с файлами с наборами

def default_serializer(obj): # для правильного преобразования даты в ISO формат
    if isinstance(obj, (date)):
        return obj.isoformat()
    raise TypeError(f'Object of type {obj.__class__.__name__} is not JSON serializable')
    

def load_cache(filename='cache.gz'): # загрузка кэша
    global CACHE
    if os.path.exists(filename) and len(CACHE)==0:
        with gzip.open(filename, 'rb') as gzip_ref:
            CACHE = json.load(gzip_ref)
        print(f'Загружено {len(CACHE)} записей из {filename} в кэш')


def save_cache(filename='cache.gz'): # сохранение кэша
    global CACHE
    with gzip.open(filename, 'wb') as gzip_file:
        json_data = json.dumps(CACHE, ensure_ascii=False, default=default_serializer)
        gzip_file.write(json_data.encode('utf-8'))
    print(f'Сохранено {len(CACHE)} записей в {filename}')
    
    
def load_gzon(filename): 
    if os.path.exists(SETS + filename + '.gz'):
        with gzip.open(SETS + filename + '.gz', 'rb') as gzip_ref:
            return json.load(gzip_ref)
    else:
        return []
    

def save_gzon(filename):
    data = globals().get(filename)
    if data is not None:
        json_data = json.dumps(data, ensure_ascii=False, default=default_serializer)
        with gzip.open(SETS + filename + '.gz', 'wb') as gzip_file:
            gzip_file.write(json_data.encode('utf-8'))
        print(f'Сохранено {len(json_data)} записей в {filename}.gz')
    else:
        print(f'Переменная {filename} не найдена')
        
     
def load_dataset(filename): 
    if os.path.exists(SETS + filename + '.gz'):
        with gzip.open(SETS + filename + '.gz', 'rb') as gzip_ref:
            return pd.DataFrame(json.load(gzip_ref))
        
    
def save_dataset(filename):
    data = globals().get(filename)
    if data is not None:
        data = data.to_dict(orient='records')
        json_data = json.dumps(data, ensure_ascii=False, default=default_serializer)
        with gzip.open(SETS + filename + '.gz', 'wb') as gzip_file:
            gzip_file.write(json_data.encode('utf-8'))
        print(f'Сохранено {len(data)} записей в {filename}.gz')
        
    
def mos_geojson(dataset):
    all_data = load_gzon(dataset)
    if(len(all_data) > 0):
        return all_data

    step = LIMIT
    url = "https://apidata.mos.ru/v1"
    folder = "/datasets/"
    params = {
        "api_key": DATA_MOS_RU,
        "$skip": 0
    }
    
    count = int(requests.get(url + folder + dataset + "/count", params=params).text)
    while count > params["$skip"]:
        response = requests.get(url + folder + dataset + "/rows", params=params)
        data = json.loads(response.text)
        for d in data:
            all_data.append(d['Cells'])
        params["$skip"] += step
    json_data = json.dumps(all_data, ensure_ascii=False)
    with gzip.open(SETS + dataset + '.gz', 'wb') as gzip_file:
        gzip_file.write(json_data.encode('utf-8'))
    print(f'Сохранено {len(all_data)} записей в {dataset}.gz')
        
    return all_data
        

def get_dataset(url, encoding='utf-8', forma='geojson')->pd.DataFrame:
    if(forma == 'geojson'):
        dataset = pd.DataFrame(mos_geojson(url))
    else:
        filename = ''
        filenames = []
        response = requests.get(url) 
        if response.status_code == 200:
            content_disposition = response.headers.get("Content-Disposition")
            if content_disposition:
                filename = re.findall("filename=(.+)", content_disposition)
                if filename:
                    filename = filename[0]
                    filename = re.sub(r'"', '', filename)
                    with open(filename, "wb") as file:
                        file.write(response.content)
                else:
                    print("Не удалось получить имя файла")
            else:
                print("Не удалось получить информацию о загруженном файле")
        else:
            print("Ошибка при загрузке файла")
        with zipfile.ZipFile(filename, 'r') as zip_ref:
            filenames = zip_ref.namelist()
        with zipfile.ZipFile(filename, 'r') as zip_ref:
            zip_ref.extract(filenames[0])
        if( forma == 'csv'):
            dataset = pd.read_csv(filenames[0], sep=';', encoding=encoding)
            dataset = dataset.drop(index=0)
            dataset = dataset.drop(columns=[col for col in dataset.columns if col.startswith("Unnamed:")])
        elif(forma == 'json'):
            dataset = pd.read_json(filenames[0], encoding=encoding)
        os.remove(filename)
        os.remove(filenames[0])
    print(dataset.info())
    return dataset

# Функция для запроса к API Яндекса
def yandex_org(row):
    global CAHCE
    load_cache()
    if pd.notna(row['geoData']):
        return row['geoData']
    
    key = f"{row['AdmArea']}_{row['District']}_{row['EDU_NAME']}"
    if key in CACHE:
        return CACHE[key]
    
    url = "https://search-maps.yandex.ru/v1/"
    params = {
        "apikey": YANDEX_ORG,
        "lang": "ru_RU",
        "text": f"Москва, {row['AdmArea']}, {row['District']}, {row['EDU_NAME']}"
    }
    
    try:
        response = requests.get(url, params=params)
        data = response.json()
        
        if 'features' in data and len(data['features']) > 0:
            geo_data = data['features'][0]['geometry']
            CACHE[key] = geo_data
            return geo_data
        else:
            CACHE[key] = None
            return None
    except Exception as e:
        print(f"Ошибка при запросе к API: {e}")
        geo_cache[key] = None
        return None

# Школы

ID:2072

## Датасет с результатами ЕГЭ

[Источник](https://data.mos.ru/opendata/2072)

[Датасет](https://data.mos.ru/odata/export/catalog?idFile=200291)

Набор данных «Результаты ЕГЭ» позволяет получить информацию об общих результатах государственной итоговой аттестации по программам среднего общего образования (далее – ГИА-11) каждой образовательной организации города Москвы.

В наборе представлены данные:

- об административных округах, районах, полных наименованиях образовательных организаций;

- о периоде, за который сдавался единый государственный экзамен (далее – ЕГЭ);

- о количестве участников, набравших не менее 200 баллов за 3 экзамена;

- о количестве участников, набравших более 160 баллов за 3 экзамена.

Результаты планируется публиковать ежегодно, таким образом можно будет проследить динамику повышения качества образования в Москве.


Современная ситуация в России

Согласно статье 59 Федерального закона от 29.12.2012 г. № 273-ФЗ «Об образовании в Российской Федерации», ГИА-11 завершает освоение основных образовательных программ среднего общего образования и является обязательной.

К ГИА-11 допускаются обучающиеся, не имеющие академической задолженности, в том числе за итоговое сочинение (изложение), и в полном объеме выполнившие учебный план или индивидуальный учебный план (имеющие годовые отметки по всем учебным предметам учебного плана за каждый год обучения по образовательной программе среднего общего образования не ниже удовлетворительных).

К ГИА-11 по учебным предметам, освоение которых завершилось ранее, допускаются обучающиеся X-XI (XII) классов, имеющие годовые отметки не ниже удовлетворительных по всем учебным предметам учебного плана за предпоследний год обучения.

Выпускники текущего года проходят ГИА-11 по двум обязательным учебным предметам (русский язык и математика) и учебным предметам по выбору.


Современная ситуация в Москве

Участники ГИА-11 могут узнать свои результаты по следующей схеме:

<table>
  <tr><td>обучающиеся образовательных организаций</td><td>в своей образовательной организации</td></tr>
  <tr><td>выпускники прошлых лет</td><td>в региональном центре обработки</td></tr>
</table>

На официальном Портале Мэра и Правительства Москвы www.mos.ru в разделе «Услуги и сервисы» можно оформить подписку на получение уведомлений о готовности результатов ГИА-11.

А знаете ли вы, что...

- Для проведения ГИА-11 предусмотрены следующие сроки: досрочный, основной и дополнительный периоды.

- Срок действия результатов ЕГЭ – 4 года, следующих за годом получения таких результатов.

- Результаты ЕГЭ каждого участника заносятся в федеральную информационную систему, бумажных свидетельств о результатах ЕГЭ не предусмотрено.


Материал подготовлен на основе информации data.mos.ru, dogm.mos.ru, mos.ru.

pd.set_option('display.max_colwidth', None)
exams = get_dataset('2072')

# чистим имена школ от мусора, делаем все буквы маленькими, убираем кавычки, и др. найденные разночтения
exams['EDU_NAME'] = exams['EDU_NAME'].str.lower().str.strip().str.replace('«', '').str.replace('»', '') \
.str.replace('  ', ' ').str.replace('"', ' ').str.replace('общеобразовательн', 'образовательн').str.replace('№ ', '')
#exams['geoData'] = None
exams.head()

## Датасет с координатами школ

ID:2263

[Источник](https://data.mos.ru/opendata/2263)

[Датасет](https://data.mos.ru/odata/export/catalog?idFile=217731)

Набор данных содержит информацию об образовательных организациях города Москвы: адреса, контактные телефоны, местоположение на карте, адреса интернет-сайтов учреждений, а также информацию о доступности зданий для лиц с ограниченными возможностями.

В соответствии с Федеральным Законом от 29.12.2012 № 273-ФЗ «Об образовании в Российской Федерации» (Закон) изменилась классификация образовательных организаций. Ранее образовательные организации классифицировались по «типам» и «видам» в зависимости от возраста обучающихся, в настоящее время согласно п.2 ст.10 Закона образование подразделяется на общее, профессиональное, дополнительное и профессиональное обучение, обеспечивающие возможность реализации права на образование в течение всей жизни (непрерывное образование).

Современные московские образовательные организации, как правило, реализуют несколько уровней общего образования (п.4 ст.10 Закона) – дошкольное, начальное общее, основное общее, среднее общее. Они также могут предлагать среднее и высшее профессиональное образование (п.5 ст.10 Закона) и четкой привязки образовательных организаций к прежней классификации (Детский сад → Средняя школа → Учреждение среднего профессионального образования → Высшее учебное заведение) уже нет.

У больших школ появились новые возможности - образовательные организации получили ресурс для повышения доступности реализуемых ими образовательных программ для московских семей. В зависимости от предпочтений семей школа предлагает дополнительное образование: организует работу кружков, секций, курсов различной направленности для детей и взрослых в любом из своих зданий.

Именно поэтому местом предоставления той или иной образовательной услуги должны считаться все здания, входящие в состав образовательной организации. Места оказания услуг общего и профессионального образования, а также – место работы конкретного кружка вы можете уточнить у администрации или на официальном сайте образовательной организации.

schools = get_dataset('2263')

schools['FullName'] = schools['FullName'].str.lower().str.strip().str.replace('«', '').str.replace('»', '') \
.str.replace('  ', ' ').str.replace('"', ' ').str.replace('общеобразовательн', 'образовательн').str.replace('№ ', '')
# Преобразование значения в колонке 'geoData' в список координат
# schools['geoData'] = schools['geoData'].str.replace('{coordinates=[', '').str.replace('], type=Polygon}', '')
schools.head()

## Координаты из списка школ в результаты ЭГЕ

exams['geoData'] = exams['EDU_NAME'].apply(lambda x: schools.loc[schools['FullName'] == x, 'geoData'].values[0] if x in schools['FullName'].values else None)
exams.head()

exams['geoData'] = exams.apply(yandex_org, axis=1)
save_cache()

exams[exams['geoData'].isna()]

# Ручное добавление одной школы"
exams.loc[exams['EDU_NAME'] == "государственное бюджетное образовательное учреждение города москвы школа 439 интеллект", 'geoData'] = "{'type': 'Point', 'coordinates': [37.745376, 55.731212]}"
exams.info()

save_dataset('exams')
